+++ packages/std/src/deps.rs
@@ -1,3 +1,7 @@
+use std::marker::PhantomData;
+
+use crate::query::CustomQuery;
+use crate::results::Empty;
 use crate::traits::{Api, Querier, Storage};
 use crate::QuerierWrapper;
 
@@ -5,27 +9,34 @@ use crate::QuerierWrapper;
 /// Designed to allow easy dependency injection at runtime.
 /// This cannot be copied or cloned since it would behave differently
 /// for mock storages and a bridge storage in the VM.
-pub struct OwnedDeps<S: Storage, A: Api, Q: Querier> {
+pub struct OwnedDeps<S: Storage, A: Api, Q: Querier, C: CustomQuery = Empty> {
     pub storage: S,
     pub api: A,
     pub querier: Q,
+    pub custom_query_type: PhantomData<C>,
 }
 
-pub struct DepsMut<'a> {
+pub struct DepsMut<'a, C: CustomQuery = Empty> {
     pub storage: &'a mut dyn Storage,
     pub api: &'a dyn Api,
-    pub querier: QuerierWrapper<'a>,
+    pub querier: QuerierWrapper<'a, C>,
 }
 
-#[derive(Copy, Clone)]
-pub struct Deps<'a> {
+#[derive(Clone)]
+pub struct Deps<'a, C: CustomQuery = Empty> {
     pub storage: &'a dyn Storage,
     pub api: &'a dyn Api,
-    pub querier: QuerierWrapper<'a>,
+    pub querier: QuerierWrapper<'a, C>,
 }
 
-impl<S: Storage, A: Api, Q: Querier> OwnedDeps<S, A, Q> {
-    pub fn as_ref(&'_ self) -> Deps<'_> {
+// Use custom implementation on order to implement Copy in case `C` is not `Copy`.
+// See "There is a small difference between the two: the derive strategy will also
+// place a Copy bound on type parameters, which isn’t always desired."
+// https://doc.rust-lang.org/std/marker/trait.Copy.html
+impl<'a, C: CustomQuery> Copy for Deps<'a, C> {}
+
+impl<S: Storage, A: Api, Q: Querier, C: CustomQuery> OwnedDeps<S, A, Q, C> {
+    pub fn as_ref(&'_ self) -> Deps<'_, C> {
         Deps {
             storage: &self.storage,
             api: &self.api,
@@ -33,7 +44,7 @@ impl<S: Storage, A: Api, Q: Querier> OwnedDeps<S, A, Q> {
         }
     }
 
-    pub fn as_mut(&'_ mut self) -> DepsMut<'_> {
+    pub fn as_mut(&'_ mut self) -> DepsMut<'_, C> {
         DepsMut {
             storage: &mut self.storage,
             api: &self.api,
@@ -42,8 +53,8 @@ impl<S: Storage, A: Api, Q: Querier> OwnedDeps<S, A, Q> {
     }
 }
 
-impl<'a> DepsMut<'a> {
-    pub fn as_ref(&'_ self) -> Deps<'_> {
+impl<'a, C: CustomQuery> DepsMut<'a, C> {
+    pub fn as_ref(&'_ self) -> Deps<'_, C> {
         Deps {
             storage: self.storage,
             api: self.api,
@@ -51,7 +62,7 @@ impl<'a> DepsMut<'a> {
         }
     }
 
-    pub fn branch(&'_ mut self) -> DepsMut<'_> {
+    pub fn branch(&'_ mut self) -> DepsMut<'_, C> {
         DepsMut {
             storage: self.storage,
             api: self.api,
@@ -63,7 +74,8 @@ impl<'a> DepsMut<'a> {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::mock::mock_dependencies;
+    use crate::mock::{mock_dependencies, MockApi, MockQuerier, MockStorage};
+    use serde::{Deserialize, Serialize};
 
     // ensure we can call these many times, eg. as sub-calls
     fn execute(mut deps: DepsMut) {
@@ -84,5 +96,35 @@ mod tests {
         let mut deps = mock_dependencies(&[]);
         execute(deps.as_mut());
         query(deps.as_ref())
+    }
+
+    #[test]
+    fn deps_implements_copy() {
+        impl CustomQuery for u64 {}
+        #[derive(Clone, Serialize, Deserialize)]
+        struct MyQuery;
+        impl CustomQuery for MyQuery {}
+
+        // With C: Copy
+        let owned = OwnedDeps::<_, _, _, u64> {
+            storage: MockStorage::default(),
+            api: MockApi::default(),
+            querier: MockQuerier::<u64>::new(&[]),
+            custom_query_type: PhantomData,
+        };
+        let deps: Deps<u64> = owned.as_ref();
+        let _copy1 = deps;
+        let _copy2 = deps;
+
+        // Without C: Copy
+        let owned = OwnedDeps::<_, _, _, MyQuery> {
+            storage: MockStorage::default(),
+            api: MockApi::default(),
+            querier: MockQuerier::<MyQuery>::new(&[]),
+            custom_query_type: PhantomData,
+        };
+        let deps: Deps<MyQuery> = owned.as_ref();
+        let _copy1 = deps;
+        let _copy2 = deps;
     }
 }