+++ packages/std/src/traits.rs
@@ -1,4 +1,5 @@
 use serde::{de::DeserializeOwned, Serialize};
+use std::marker::PhantomData;
 use std::ops::Deref;
 
 use crate::addresses::{Addr, CanonicalAddr};
@@ -134,41 +135,42 @@ pub trait Querier {
     fn raw_query(&self, bin_request: &[u8]) -> QuerierResult;
 }
 
-#[derive(Copy, Clone)]
-pub struct QuerierWrapper<'a>(&'a dyn Querier);
+#[derive(Clone)]
+pub struct QuerierWrapper<'a, C: CustomQuery = Empty> {
+    querier: &'a dyn Querier,
+    custom_query_type: PhantomData<C>,
+}
+
+// Use custom implementation on order to implement Copy in case `C` is not `Copy`.
+// See "There is a small difference between the two: the derive strategy will also
+// place a Copy bound on type parameters, which isn’t always desired."
+// https://doc.rust-lang.org/std/marker/trait.Copy.html
+impl<'a, C: CustomQuery> Copy for QuerierWrapper<'a, C> {}
 
 /// This allows us to use self.raw_query to access the querier.
 /// It also allows external callers to access the querier easily.
-impl<'a> Deref for QuerierWrapper<'a> {
+impl<'a, C: CustomQuery> Deref for QuerierWrapper<'a, C> {
     type Target = dyn Querier + 'a;
 
     fn deref(&self) -> &Self::Target {
-        self.0
+        self.querier
     }
 }
 
-impl<'a> QuerierWrapper<'a> {
+impl<'a, C: CustomQuery> QuerierWrapper<'a, C> {
     pub fn new(querier: &'a dyn Querier) -> Self {
-        QuerierWrapper(querier)
+        QuerierWrapper {
+            querier,
+            custom_query_type: PhantomData,
+        }
     }
 
-    /// query is a shorthand for custom_query when we are not using a custom type,
-    /// this allows us to avoid specifying "Empty" in all the type definitions.
-    pub fn query<T: DeserializeOwned>(&self, request: &QueryRequest<Empty>) -> StdResult<T> {
-        self.custom_query(request)
-    }
-
-    /// Makes the query and parses the response. Also handles custom queries,
-    /// so you need to specify the custom query type in the function parameters.
-    /// If you are no using a custom query, just use `query` for easier interface.
+    /// Makes the query and parses the response.
     ///
     /// Any error (System Error, Error or called contract, or Parse Error) are flattened into
     /// one level. Only use this if you don't need to check the SystemError
     /// eg. If you don't differentiate between contract missing and contract returned error
-    pub fn custom_query<C: CustomQuery, U: DeserializeOwned>(
-        &self,
-        request: &QueryRequest<C>,
-    ) -> StdResult<U> {
+    pub fn query<U: DeserializeOwned>(&self, request: &QueryRequest<C>) -> StdResult<U> {
         let raw = to_vec(request).map_err(|serialize_err| {
             StdError::generic_err(format!("Serializing QueryRequest: {}", serialize_err))
         })?;
@@ -330,7 +332,7 @@ mod tests {
     #[test]
     fn use_querier_wrapper_as_querier() {
         let querier: MockQuerier<Empty> = MockQuerier::new(&[]);
-        let wrapper = QuerierWrapper::new(&querier);
+        let wrapper = QuerierWrapper::<Empty>::new(&querier);
 
         // call with deref shortcut
         let res = demo_helper(&*wrapper);
@@ -345,7 +347,7 @@ mod tests {
     fn auto_deref_raw_query() {
         let acct = String::from("foobar");
         let querier: MockQuerier<Empty> = MockQuerier::new(&[(&acct, &coins(5, "BTC"))]);
-        let wrapper = QuerierWrapper::new(&querier);
+        let wrapper = QuerierWrapper::<Empty>::new(&querier);
         let query = QueryRequest::<Empty>::Bank(BankQuery::Balance {
             address: acct,
             denom: "BTC".to_string(),